+++ src/Links.php
@@ -134,10 +134,10 @@ class Links implements ArrayAccess, Arrayable, Jsonable, JsonSerializable
     /**
      * {@inheritdoc}
      *
-     * @return array|mixed
+     * @return object
      */
     public function jsonSerialize()
     {
-        return $this->toArray();
+        return (object) $this->toArray();
     }
 }